// src/app/models/thong-ke-nhiem-vu/bieu-do-thong-ke-theo-tieu-chi.ts
export interface BieuDoThongKeTheoTieuChi {}

export interface BieuDoTheoTieuChiModel {
    hoanThanhDungHan?: any[];
    hoanThanhDungHanCount?: string;
    dangHoanThanhDungHan?: any[];
    dangHoanThanhDungHanCount?: string;
    hoanThanhQuaHan?: any[];
    hoanThanhQuaHanCount?: string;
    dangHoanThanhQuaHan?: any[];
    dangHoanThanhQuaHanCount?: string;
    label?: any[];
    value?: number;
    count?: number;
}